--- shapiq/explainer/tree/__init__.py
@@ -1,4 +1,0 @@
-"""This module contains the tree explainer implementation."""
-from .treeshap_iq import TreeExplainer
-
-__all__ = ["TreeExplainer"]
