+++ shapiq/explainer/treeshap_iq/__init__.py
@@ -1,0 +1,5 @@
+"""This module contains the tree explainer implementation."""
+from .conversion import TreeModel
+from .treeshap_iq import TreeSHAPIQ
+
+__all__ = ["TreeSHAPIQ", "TreeModel"]
