+++ shapiq/explainer/treeshap_iq/conversion/__init__.py
@@ -1,6 +1,6 @@
 """This module collects the conversion functions for the tree explainer implementation."""
-from .base import TreeModel, EdgeTree
-from .sklearn import convert_sklearn_tree
+from .base import EdgeTree, TreeModel
 from .edge_representation import create_edge_tree
+from .sklearn import convert_sklearn_tree
 
 __all__ = ["convert_sklearn_tree", "TreeModel", "EdgeTree", "create_edge_tree"]
